Rename UpCommand to Up
